renamed usercount => stats
