+++ keycloak-stats-spi/src/main/resources/jboss-deployment-structure.xml
