+++ pom.xml
@@ -16,7 +16,7 @@
     <modules>
         <module>bcrypt-addon</module>
         <module>jsonlog-eventlistener</module>
-        <module>keycloak-metrics-spi</module>
+        <module>keycloak-stats-spi</module>
 <!--        <module>user-removed-eventlistener</module>-->
     </modules>
